# Remove unnecessary unaliasing allocation