@GunnarFarneback 2024-09-23
Looks good. Notice that this was one of the optimizations identified in #9, but either this implementation is better or it has become more effective since then.